+++ src/client/modules/graphic/components/orchard-logo/graphic-orchard-logo.component.html
@@ -1,41 +1,22 @@
 <svg id="graphic-orchard-logo" data-name="graphic-orchard-logo" [attr.height]="height" viewBox="0 0 190 190">
-	<defs>
-		<style>
-		.cls-1 {
-			fill: #ffd61f;
-		}
-
-		.cls-2 {
-			fill: #100;
-		}
-
-		.cls-3 {
-			fill: #f3651d;
-		}
-
-		.cls-4 {
-			fill: #fffd9f;
-		}
-
-		.cls-5 {
-			fill: #9c2222;
-		}
-
-		.cls-6 {
-			fill: #f58f22;
-		}
-		</style>
-	</defs>
-	<rect class="cls-2" x="0" y="0" width="190" height="190" rx="95" ry="95"/>
+	<rect *ngIf="surface" class="orc-brand-surface-fill" x="0" y="0" width="190" height="190" rx="95" ry="95"/>
 	<g>
-		<path class="cls-6" d="M182.57,83.82l-57.57,34.86c-7.95-2.63-16.57-4.14-24.56-5.4l78.01-47.24c1.98,5.75,3.36,11.71,4.12,17.78Z"/>
-		<path class="cls-1" d="M20.46,47.53c.96-1.52,2.63-2.35,4.33-2.35.91,0,1.83.24,2.65.74l58.31,35.31c-3.61,1.72-8.79,4.33-14.08,7.49-1.23.73-2.38,1.46-3.5,2.18L14.56,58.43c1.7-3.75,3.67-7.4,5.9-10.91Z"/>
-		<path class="cls-5" d="M36.99,161.35c-11.18-9.76-19.58-22.05-24.66-35.63l46.61,28.22c-7.49,3.07-15.68,5.63-21.95,7.41Z"/>
-		<path class="cls-3" d="M72.4,131.54c.98.33,1.82.61,2.28.8,1.07.44,1.93,1.06,2.48,1.81,1.5,2.01,2.36,3.78,1.65,5.9-1.18,3.48-5.49,6.82-11.17,9.85l.02-.03L8.91,114.3c-1.43-6.34-2.18-12.86-2.18-19.47,0-.71.02-1.43.04-2.14l46.46,28.13c4.88,5.9,14.29,9.08,19.17,10.72Z"/>
-		<path class="cls-6" d="M50.46,109.79L7.44,83.74c.76-6.07,2.15-12.02,4.14-17.77l49.49,29.97c-6.07,4.85-9.57,9.43-10.6,13.85Z"/>
-		<path class="cls-1" d="M71.58,103.53c-.04-1.35,1.81-3.28,3.68-4.67l87.34-52.89c2.37-1.44,5.51-.71,6.99,1.61,2.22,3.5,4.19,7.15,5.88,10.9l-87.24,52.83c-4.49-.81-8.37-1.71-10.88-2.9-4.21-2-5.74-3.83-5.77-4.89Z"/>
-		<path class="cls-3" d="M134.19,122.47l49.05-29.7c.02.69.04,1.37.04,2.06,0,6.65-.76,13.21-2.21,19.57l-32.82,19.87c-.16-.23-.31-.46-.47-.68-3.48-4.78-8.2-8.36-13.59-11.12Z"/>
-		<path class="cls-5" d="M153.11,161.27c1.18-7.43.76-14.07-1.25-19.84l25.78-15.61c-5.07,13.5-13.42,25.72-24.53,35.44Z"/>
-		<path class="cls-4" d="M71.69,46.15c-1.22,4.4-2.31,16.68,20.69,22.58v10.14l2.64-1.4,2.54-1.54v-7.21c22.46-5.77,21.95-17.61,20.77-22.27-.17-.66-.44-1.28-.79-1.86l-9.23-15.19c-8.82,8.72-16.17,3.37-16.17,3.37,7.3.93,11.07-4.41,12.78-8.15-1.5-1.69-3.11-2.99-4.78-3.88-.04-.02-.07-.03-.11-.05-.31-.16-.61-.29-.92-.42-.17-.07-.34-.14-.51-.2-.14-.05-.27-.09-.41-.13-1.05-.33-2.11-.51-3.18-.51-4.7,0-9.27,3.26-12.87,9.17l-.51.85c-.89,2.01-.91,4.32.06,6.32,3.31,6.86,9.76,15.54,21.43,16.77,0,0-13.23,4.24-27.97-12.45l-3.17,5.21c-.11.27-.21.54-.29.82Z"/>
+		<path class="orc-brand-lightest-fill" d="M71.69,46.15c-1.22,4.4-2.31,16.68,20.69,22.58v10.14l2.64-1.4,2.54-1.54v-7.21c22.46-5.77,21.95-17.61,20.77-22.27-.17-.66-.44-1.28-.79-1.86l-9.23-15.19c-8.82,8.72-16.17,3.37-16.17,3.37,7.3.93,11.07-4.41,12.78-8.15-1.5-1.69-3.11-2.99-4.78-3.88-.04-.02-.07-.03-.11-.05-.31-.16-.61-.29-.92-.42-.17-.07-.34-.14-.51-.2-.14-.05-.27-.09-.41-.13-1.05-.33-2.11-.51-3.18-.51-4.7,0-9.27,3.26-12.87,9.17l-.51.85c-.89,2.01-.91,4.32.06,6.32,3.31,6.86,9.76,15.54,21.43,16.77,0,0-13.23,4.24-27.97-12.45l-3.17,5.21c-.11.27-.21.54-.29.82Z"/>
+		<g class="orc-brand-light-fill">
+			<path d="M20.46,47.53c.96-1.52,2.63-2.35,4.33-2.35.91,0,1.83.24,2.65.74l58.31,35.31c-3.61,1.72-8.79,4.33-14.08,7.49-1.23.73-2.38,1.46-3.5,2.18L14.56,58.43c1.7-3.75,3.67-7.4,5.9-10.91Z"/>
+			<path d="M71.58,103.53c-.04-1.35,1.81-3.28,3.68-4.67l87.34-52.89c2.37-1.44,5.51-.71,6.99,1.61,2.22,3.5,4.19,7.15,5.88,10.9l-87.24,52.83c-4.49-.81-8.37-1.71-10.88-2.9-4.21-2-5.74-3.83-5.77-4.89Z"/>
+		</g>
+		<g class="orc-brand-medium-fill">
+			<path d="M182.57,83.82l-57.57,34.86c-7.95-2.63-16.57-4.14-24.56-5.4l78.01-47.24c1.98,5.75,3.36,11.71,4.12,17.78Z"/>
+			<path d="M50.46,109.79L7.44,83.74c.76-6.07,2.15-12.02,4.14-17.77l49.49,29.97c-6.07,4.85-9.57,9.43-10.6,13.85Z"/>
+		</g>
+		<g class="orc-brand-dark-fill">
+			<path d="M72.4,131.54c.98.33,1.82.61,2.28.8,1.07.44,1.93,1.06,2.48,1.81,1.5,2.01,2.36,3.78,1.65,5.9-1.18,3.48-5.49,6.82-11.17,9.85l.02-.03L8.91,114.3c-1.43-6.34-2.18-12.86-2.18-19.47,0-.71.02-1.43.04-2.14l46.46,28.13c4.88,5.9,14.29,9.08,19.17,10.72Z"/>
+			<path d="M134.19,122.47l49.05-29.7c.02.69.04,1.37.04,2.06,0,6.65-.76,13.21-2.21,19.57l-32.82,19.87c-.16-.23-.31-.46-.47-.68-3.48-4.78-8.2-8.36-13.59-11.12Z"/>
+		</g>
+		<g class="orc-brand-darkest-fill">
+			<path class="cls-5" d="M36.99,161.35c-11.18-9.76-19.58-22.05-24.66-35.63l46.61,28.22c-7.49,3.07-15.68,5.63-21.95,7.41Z"/>
+			<path class="cls-5" d="M153.11,161.27c1.18-7.43.76-14.07-1.25-19.84l25.78-15.61c-5.07,13.5-13.42,25.72-24.53,35.44Z"/>
+		</g>
 	</g>
 </svg>
