+++ src/client/modules/nav/components/primary-nav/primary-nav.component.html
@@ -1,0 +1,9 @@
+<div class="primary-nav-container">
+    <div class="primary-nav-header">
+		<orc-graphic-orchard-logo height="3.5rem"></orc-graphic-orchard-logo>
+		<div class="primary-nav-epoch">Epoch 5</div>
+		<div class="primary-nav-blockheight">844345</div>
+	</div>
+    <div class="primary-nav-items">nav items</div>
+    <div>footer</div>
+</div>
