--- src/client/modules/nav/components/primary/primary-nav.component.html
@@ -1,7 +1,0 @@
-<div class="primary-nav-container">
-    <div>
-		<orc-graphic-orchard-logo></orc-graphic-orchard-logo>
-	</div>
-    <div class="primary-nav-items">nav items</div>
-    <div>footer</div>
-</div>
