--- src/client/modules/nav/components/primary/primary-nav.component.scss
@@ -1,13 +1,0 @@
-@use '@styles/molecule/molecule';
-
-.primary-nav-container{
-    @extend .orc-surface-container-low-bg;
-    @extend .flex;
-    @extend .flex-column;
-    @extend .flex-items-center;
-    @extend .h-full;
-}
-
-.primary-nav-items{
-    @extend .flex-grow;
-}
