--- src/client/modules/nav/components/primary/primary-nav.component.ts
@@ -1,12 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'orc-primary-nav',
-  standalone: false,
-  templateUrl: './primary-nav.component.html',
-  styleUrl: './primary-nav.component.scss',
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class PrimaryNavComponent {
-
-}
