+++ src/client/modules/routing/components/layout-interior/layout-interior.component.html
@@ -1,11 +1,3 @@
-<!-- <div class="menu">
-  some menu
-</div>
-<div class="main-content">
-  <router-outlet></router-outlet>
-</div>
- -->
-
 <div class="interior-surface">
   <orc-primary-nav></orc-primary-nav>
   <div class="main-content">
