+++ src/client/styles/molecule/_molecule.scss
@@ -1,2 +1,4 @@
 @use './token';
+@use './theme';
 @use './position';
+@use './typetext';
