+++ src/client/styles/molecule/position.scss
@@ -28,4 +28,48 @@
     width: 100%;
 }
 
+/* Minimum Width Definitions */
+.w-min-7{
+    min-width: 7rem;
+}
+
 /* Margin Definitions */
+
+/* Padding Definitions */
+.p-v-1-5{
+    padding-top: 1.5rem;
+    padding-bottom: 1.5rem;
+}
+
+.p-b-0-5{
+    padding-bottom: 0.5rem;
+}
+
+// .p-t-1{
+//     padding-top: 1rem;
+// }
+// .p-t-2{
+//     padding-top: 2rem;
+// }
+// .p-1{
+//     padding: 1rem;
+// }
+// .p-2{
+//     padding: 2rem;
+// }
+// .p-v-1{
+//     padding-top: 1rem;
+//     padding-bottom: 1rem;
+// }
+// .p-v-2{
+//     padding-top: 2rem;
+//     padding-bottom: 2rem;
+// }
+// .p-h-1{
+//     padding-left: 1rem;
+//     padding-right: 1rem;
+// }
+// .p-h-2{
+//     padding-left: 2rem;
+//     padding-right: 2rem;
+// }
