+++ src/client/styles/molecule/typetext.scss
@@ -1,0 +1,27 @@
+/* Font Weight Definitions */
+.font-weight-extra-light{
+    font-weight: 200;
+}
+.font-weight-light{
+    font-weight: 300;
+}
+.font-weight-regular{
+    font-weight: 400;
+}
+.font-weight-medium{
+    font-weight: 500;
+}
+.font-weight-semi-bold{
+    font-weight: 600;
+}
+.font-weight-bold{
+    font-weight: 700;
+}
+
+/* Font Size Definitions */
+.font-size-xs {
+    font-size: 0.7rem;
+}
+.font-size-s {
+    font-size: 0.8rem;
+}
